Persistencia con postgres

+++ tiendaSolis/src/main/java/com/tiendaSolis/restful/tiendaSolis/repository/ProductRepository.java
@@ -15,6 +15,6 @@ import org.springframework.stereotype.Repository;
  * @author julio
  */
 @Repository
-public interface ProductRepository extends JpaRepository<Product, BigInteger>{
+public interface ProductRepository extends JpaRepository<Product, Integer>{
     
 }
